third_party/WebKit/LayoutTests/fast/css/invalidation/disabled-pseudo.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
input { background: transparent }
:disabled { background-color: green }
input + div { color: pink }
</style>
<input id="button" type="button"></input>
<div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
description("Use descendant invalidation set for :disabled pseudo class.")

var transparent = "rgba(0, 0, 0, 0)";
var green = "rgb(0, 128, 0)";

shouldBe("getComputedStyle(button, '').backgroundColor", "transparent");

button.offsetTop; // Force recalc.

button.disabled = true;

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

shouldBe("getComputedStyle(button, '').backgroundColor", "green");
</script>
